emacs: TRAMP fixes.
There were two issues I was running into. The first has to do with
projectile which can be solved by turning off caching:

https://github.com/bbatsov/prelude/issues/594

The second was strange. If "." is in the load-path then TRAMP hangs on
trying to run sudo, at "Sending password" (it does not even ask for a
password).